+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v1.8.4
+
+### Changed
+
+- Changed the cache key and TTL parameter names in the View class to be more concise (`$key` and `$ttl`).
+- Add `post_mime_type` to the attachment factory definition to ensure that
+  attachments are created with a valid MIME type.
 
 ### Fixed